Configure JWT Authentication

+++ backend/controllers/expenseController.js
@@ -1,12 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Expense = require('../models/expenseModel')
+const User = require('../models/userModel')
 
 // @desc Get expenses
 // @route GET /api/expenses
 // @access Private
 const getExpenses = asyncHandler(async (req, res) => {
-    const expenses = await Expense.find()
+    const expenses = await Expense.find({ user: req.user.id})
+    
     res.status(200).json(expenses)
 })
 
@@ -23,6 +25,7 @@ const setExpense = asyncHandler(async (req, res) => {
         amount: req.body.amount,
         description: req.body.description,
         date: req.body.date, 
+        user: req.user.id,
     })
     res.status(200).json(expense)
 })
@@ -37,6 +40,20 @@ const updateExpense = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Expense not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure logged in user matches the expense user
+    if(expense.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedExpense = await Expense.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
     })
@@ -53,6 +70,20 @@ const deleteExpense = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Expense not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure logged in user matches the expense user
+    if(expense.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await expense.deleteOne()
     res.status(200).json({ id: req.params.id })
 })
